Move `Config` and associated `const`s to `config.rs`
... as we're about to add a bunch more stuff, it improves readability to
split these out to their own module.
